Create common core for 0.13 and 1.0.
For now both 1.0 and 0.13 it's own Engine but share rest of configuration (that is declared in 0.13).
