+++ 0.13/hoarder/src/main/sbt_0.13/org.romanowski.hoarder/core/SbtAnalysisMapper.scala
@@ -66,6 +66,3 @@ class SbtAnalysisMapper(sbtOutput: Path,
   override val productStampMapper: ContextAwareMapper[File, Stamp] =
     Mapper.updateModificationDateFileMapper(productMapper)
 }
-
-
-
