+++ 0.13/hoarder/src/main/sbt_0.13/sbt/inc/MappableFormat.scala
