+++ 0.13/hoarder/src/test/sbt_0.13/sbt/inc/TextAnalysisFormatSpecification.scala
