+++ 1.0/project/build.properties
@@ -1,1 +1,1 @@
-sbt.version = 1.0.0-cached-SNAPSHOT
+sbt.version = 1.0.0-M5
